app/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
motor>=3.3
pymongo>=4.6
pydantic>=2.6
python-dotenv>=1.0
#sentence-transformers>=3.0
rank-bm25>=0.2.2
numpy>=1.24
requests>=2.31
PyJWT>=2.8
